:sparkles: Front: load JSON test random terrain into MapRenderer story
