+++ micropolis_web_front/package.json
@@ -44,7 +44,7 @@
         "babel-loader": "^8.2.2",
         "babel-plugin-macros": "^3.0.1",
         "babel-preset-react-app": "^10.0.0",
-        "copy-webpack-plugin": "^7.0.0",
+        "copy-webpack-plugin": "^6.4.1",
         "cssnano": "^4.1.10",
         "eslint": "^7.19.0",
         "eslint-config-airbnb-typescript": "^12.0.0",
